feat: change publication links for 2021 and 2022
There are multiple versions of the 2021 and 2022 reports.  When linking from the publication page send users to the appropriate report page where they can select their language of choice.  This is an extra click but is preferred over simply linking direcly to the English-language PDF.

The link text is also altered from "Read PDF" to "Download the report" to more clearly indicate what will happen next.

Signed-off-by: Nathen Harvey <nathenharvey@google.com>