A test framework for the redirects
This builds out a script to validate redirects for the application

`redirects.csv` is a comma separated list of redirects:  source,
  destination, status code

`test-redirects.js` loads the csv file and validates that each redirect
works correctly.

A `BASE_URL` environment variable can be used to set the base url,
`https://dora.dev` is used by default.

A `Dockerfile` is included so that this can be packaged up into a
container.

Signed-off-by: Nathen Harvey <nathenharvey@google.com>