Test: Improve version checker tests with data attributes
Refactor version checker tests to use data-version attributes for more robust and maintainable selectors. This approach decouples the tests from the visual presentation of the page and makes them less prone to breaking with HTML structure changes.

The isFirst property is removed from the versions array as it's no longer needed with the new selectors. Assertions are also simplified to focus on the presence or absence of the target element.

Signed-off-by: Nathen Harvey <nathenharvey@google.com>