Add measurement frameworks research article
Adds a new article on choosing measurement frameworks to fit organizational goals.

This article is both an AI research article and a 2025 research article.

This change uses hugo's headless bundle feature along with a shortcode to render the bundle in multiple places. Doing so means that the article only needs to be maintained in one place. There is a tradeoff: the front matter needs to be maintained in multiple places.

Alternate options considered:

* Pick just one place for the content
* Elevate this article to a guide


This includes:
 - Two new research pages (under /2025 and /ai)
   - Shared content for the article.
 - Tests are included for the two new pages.
 - New SVG images to support the article content.
 - Updated theme with figcaption styles.
 - Links to the new article from the 2025 and AI research collections.